Split FlexRole, FlexRoleManager and FlexActors into different files. Did the same with the tests files.
