+++ flex/actors/__init__.py
@@ -2,6 +2,6 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import absolute_import
 
-from flex.actors.actors import FlexRole
-from flex.actors.actors import FlexRoleManager
+from flex.actors.role import FlexRole
+from flex.actors.role_manager import FlexRoleManager
 from flex.actors.actors import FlexActors
