+++ tests/actors/test_actors.py
@@ -2,7 +2,8 @@ import unittest
 
 import pytest
 
-from flex.actors.actors import FlexActors, FlexRole, FlexRoleManager
+from flex.actors.actors import FlexActors
+from flex.actors.role import FlexRole
 
 
 @pytest.fixture(name="actors_cs")
@@ -49,164 +50,6 @@ def fixture_peer_to_peer():
             actor5_id: role5,
         }
     )
-
-
-class TestRoleManger(unittest.TestCase):
-    @pytest.fixture(autouse=True)
-    def _fixture_simple_actors_client_server(self, actors_cs):
-        self._actors_cs = actors_cs
-
-    @pytest.fixture(autouse=True)
-    def _fixture_peer_to_peer(self, p2p):
-        self._p2p = p2p
-
-    def test_is_client_method(self):
-        # Client-Server Architecture
-        assert FlexRoleManager.is_client(self._actors_cs["client1"]) is True
-        assert FlexRoleManager.is_client(self._actors_cs["client2"]) is True
-        assert FlexRoleManager.is_client(self._actors_cs["client3"]) is True
-        assert FlexRoleManager.is_client(self._actors_cs["aggregator"]) is False
-        assert FlexRoleManager.is_client(self._actors_cs["server"]) is False
-        # Peer-To-Peer Architecture
-        assert FlexRoleManager.is_client(self._p2p["actor1"]) is True
-        assert FlexRoleManager.is_client(self._p2p["actor2"]) is True
-        assert FlexRoleManager.is_client(self._p2p["actor3"]) is True
-        assert FlexRoleManager.is_client(self._p2p["actor4"]) is True
-        assert FlexRoleManager.is_client(self._p2p["actor5"]) is True
-
-    def test_is_aggregator_method(self):
-        # Client-Server Architecture
-        assert FlexRoleManager.is_aggregator(self._actors_cs["client1"]) is False
-        assert FlexRoleManager.is_aggregator(self._actors_cs["client2"]) is False
-        assert FlexRoleManager.is_aggregator(self._actors_cs["client3"]) is False
-        assert FlexRoleManager.is_aggregator(self._actors_cs["aggregator"]) is True
-        assert FlexRoleManager.is_aggregator(self._actors_cs["server"]) is False
-        # Peer-To-Peer Architecture
-        assert FlexRoleManager.is_aggregator(self._p2p["actor1"]) is True
-        assert FlexRoleManager.is_aggregator(self._p2p["actor2"]) is True
-        assert FlexRoleManager.is_aggregator(self._p2p["actor3"]) is True
-        assert FlexRoleManager.is_aggregator(self._p2p["actor4"]) is True
-        assert FlexRoleManager.is_aggregator(self._p2p["actor5"]) is True
-
-    def test_is_server_method(self):
-        # Client-Server Architecture
-        assert FlexRoleManager.is_server(self._actors_cs["client1"]) is False
-        assert FlexRoleManager.is_server(self._actors_cs["client2"]) is False
-        assert FlexRoleManager.is_server(self._actors_cs["client3"]) is False
-        assert FlexRoleManager.is_server(self._actors_cs["aggregator"]) is False
-        assert FlexRoleManager.is_server(self._actors_cs["server"]) is True
-        # Peer-To-Peer Architecture
-        assert FlexRoleManager.is_server(self._p2p["actor1"]) is True
-        assert FlexRoleManager.is_server(self._p2p["actor2"]) is True
-        assert FlexRoleManager.is_server(self._p2p["actor3"]) is True
-        assert FlexRoleManager.is_server(self._p2p["actor4"]) is True
-        assert FlexRoleManager.is_server(self._p2p["actor5"]) is True
-
-    def test_check_compatibility_method(self):
-        # Client-Server Architecture
-        assert (
-            FlexRoleManager.check_compatibility(
-                self._actors_cs["client1"], self._actors_cs["server"]
-            )
-            is False
-        )
-        assert (
-            FlexRoleManager.check_compatibility(
-                self._actors_cs["server"], self._actors_cs["client1"]
-            )
-            is True
-        )
-        assert (
-            FlexRoleManager.check_compatibility(
-                self._actors_cs["server"], self._actors_cs["aggregator"]
-            )
-            is False
-        )
-        assert (
-            FlexRoleManager.check_compatibility(
-                self._actors_cs["client1"], self._actors_cs["client2"]
-            )
-            is False
-        )
-        assert (
-            FlexRoleManager.check_compatibility(
-                self._actors_cs["aggregator"], self._actors_cs["server"]
-            )
-            is True
-        )
-        assert (
-            FlexRoleManager.check_compatibility(
-                self._actors_cs["client1"], self._actors_cs["aggregator"]
-            )
-            is True
-        )
-        # Peer-To-Peer Architecture
-        assert (
-            FlexRoleManager.check_compatibility(
-                self._p2p["actor1"], self._p2p["actor2"]
-            )
-            is True
-        )
-        assert (
-            FlexRoleManager.check_compatibility(
-                self._p2p["actor1"], self._p2p["actor3"]
-            )
-            is True
-        )
-        assert (
-            FlexRoleManager.check_compatibility(
-                self._p2p["actor1"], self._p2p["actor4"]
-            )
-            is True
-        )
-        assert (
-            FlexRoleManager.check_compatibility(
-                self._p2p["actor1"], self._p2p["actor5"]
-            )
-            is True
-        )
-        assert (
-            FlexRoleManager.check_compatibility(
-                self._p2p["actor2"], self._p2p["actor3"]
-            )
-            is True
-        )
-        assert (
-            FlexRoleManager.check_compatibility(
-                self._p2p["actor2"], self._p2p["actor4"]
-            )
-            is True
-        )
-        assert (
-            FlexRoleManager.check_compatibility(
-                self._p2p["actor2"], self._p2p["actor5"]
-            )
-            is True
-        )
-        assert (
-            FlexRoleManager.check_compatibility(
-                self._p2p["actor3"], self._p2p["actor4"]
-            )
-            is True
-        )
-        assert (
-            FlexRoleManager.check_compatibility(
-                self._p2p["actor3"], self._p2p["actor5"]
-            )
-            is True
-        )
-        assert (
-            FlexRoleManager.check_compatibility(
-                self._p2p["actor4"], self._p2p["actor5"]
-            )
-            is True
-        )
-        assert (
-            FlexRoleManager.check_compatibility(
-                self._p2p["actor5"], self._p2p["actor3"]
-            )
-            is True
-        )
 
 
 class TestActors(unittest.TestCase):
